+++ src/datasources/redis/components/RedisQueryEditor.tsx
@@ -35,7 +35,7 @@ export class RedisQueryEditor extends PureComponent<Props, State> {
             legendFormat: query.legendFormat,
             options: {
                 rateConversion: query.options.rateConversion,
-                timeUtilizationConversion: query.options.rateConversion,
+                timeUtilizationConversion: query.options.timeUtilizationConversion,
             },
         };
     }
@@ -53,6 +53,11 @@ export class RedisQueryEditor extends PureComponent<Props, State> {
         this.setState({ format }, this.runQuery);
     };
 
+    onRateConversionChange = (event: React.SyntheticEvent<HTMLInputElement>) => {
+        const rateConversion = (event.target as HTMLInputElement).checked;
+        this.setState({ options: { ...this.state.options, rateConversion } }, this.runQuery);
+    };
+
     onTimeUtilizationConversionChange = (event: React.SyntheticEvent<HTMLInputElement>) => {
         const timeUtilizationConversion = (event.target as HTMLInputElement).checked;
         this.setState({ options: { ...this.state.options, timeUtilizationConversion } }, this.runQuery);
@@ -66,7 +71,7 @@ export class RedisQueryEditor extends PureComponent<Props, State> {
             legendFormat: this.state.legendFormat,
             options: {
                 rateConversion: this.state.options.rateConversion,
-                timeUtilizationConversion: this.state.options.rateConversion,
+                timeUtilizationConversion: this.state.options.timeUtilizationConversion,
             },
         };
     };
@@ -128,6 +133,23 @@ export class RedisQueryEditor extends PureComponent<Props, State> {
                             options={FORMAT_OPTIONS}
                             value={this.state.format}
                             onChange={this.onFormatChange}
+                        />
+                    </InlineField>
+
+                    <InlineField label="Rate Conversion" tooltip="Counter metrics will be converted to a rate">
+                        <InlineSwitch
+                            value={this.state.options.rateConversion}
+                            onChange={this.onRateConversionChange}
+                        />
+                    </InlineField>
+
+                    <InlineField
+                        label="Time Utilization Conversion"
+                        tooltip="Time-based counter metrics will be converted to a utilization (for example kernel.all.cpu.user)"
+                    >
+                        <InlineSwitch
+                            value={this.state.options.timeUtilizationConversion}
+                            onChange={this.onTimeUtilizationConversionChange}
                         />
                     </InlineField>
                 </InlineFieldRow>