+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Enhancements / Bug Fixes
 
+- **redis, vector**: add buttons to disable rate conversation and time utilization conversation
 - **redis**: use LRU cache for series metadata
 - **redis**: fix label_names() function to return all label names if no parameter is specified (now the label name auto-completion in the query editor works again)
 - **redis**: remove deprecated `label_values(metric, label)` function